Move Dao and Database to Package
