+++ app/src/main/java/com/capstone/project/tourify/data/local/room/detail/DetailDao.kt
@@ -1,4 +1,4 @@
-package com.capstone.project.tourify.data.local.room
+package com.capstone.project.tourify.data.local.room.detail
 
 import androidx.room.Dao
 import androidx.room.Insert
