+++ app/src/main/java/com/capstone/project/tourify/data/repository/UserRepository.kt
@@ -11,8 +11,8 @@ import com.capstone.project.tourify.data.remote.response.ArticlesResponseItem
 import com.capstone.project.tourify.data.remote.retrofit.ApiService
 import com.dicoding.picodiploma.loginwithanimation.data.ArticleRemoteMediator
 import com.capstone.project.tourify.data.local.entity.CategoryEntity
-import com.capstone.project.tourify.data.local.room.CategoryDao
-import com.capstone.project.tourify.data.local.room.DetailDao
+import com.capstone.project.tourify.data.local.room.category.CategoryDao
+import com.capstone.project.tourify.data.local.room.detail.DetailDao
 import com.capstone.project.tourify.data.remote.response.CategoryResponseItem
 import com.capstone.project.tourify.data.remote.response.DetailResponse
